% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_prob.R
\name{select_prob}
\alias{select_prob}
\title{Probabilities to select a normalization set without DE-gene}
\usage{
select_prob(m, k, invariant = TRUE)
}
\arguments{
\item{m, }{number of genes}

\item{k, }{normalization subset size}

\item{invariant, }{boolean, when TRUE, probability of selection is evaluated for invariant gene}
}
\value{
a vector of probabilities of having at least one differential expression used as an reference selected in the normalization subset for any number of differential expressions d in the gene collection.
}
\description{
Probabilities to select a normalization set without DE-gene
}
\examples{

select_prob(500, 10, invariant=TRUE)
}
